requirements.txt
neo4j==5.28.2
numpy==2.3.3
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
fastapi==0.117.1
uvicorn==0.32.1
python-dotenv==1.0.1
openai==1.57.2
pydantic==2.10.4
